Fold Link Migration into Work Item migration (#292)
* Added Link migration directly to Work item Migrator so you only need 1 run. the Link Migrator still exists and both use the same code.

* Added LinkMigration switch to Work Item Migrator

* removed random file.

* |Updated Version
