+++ GitVersion.yml
@@ -1,7 +1,7 @@
 assembly-versioning-scheme: MajorMinorPatch
 mode: ContinuousDeployment
 continuous-delivery-fallback-tag: ''
-next-version: 8.0.0
+next-version: 8.1.0
 branches:
   master:
     mode: ContinuousDeployment
