+++ src/VstsSyncMigrator.Core/Configuration/EngineConfiguration.cs
@@ -114,7 +114,7 @@ namespace VstsSyncMigrator.Engine.Configuration
                     { "Product Backlog Item", "Product Backlog Item" }
             };
             ec.Processors = new List<ITfsProcessingConfig>();
-            ec.Processors.Add(new WorkItemMigrationConfig() { Enabled = false, ReplayRevisions= true, UpdateCreatedBy =true, PrefixProjectToNodes = false, UpdateCreatedDate=true, UpdateSourceReflectedId=true, QueryBit = @"AND [TfsMigrationTool.ReflectedWorkItemId] = '' AND  [Microsoft.VSTS.Common.ClosedDate] = '' AND [System.WorkItemType] IN ('Shared Steps', 'Shared Parameter', 'Test Case', 'Requirement', 'Task', 'User Story', 'Bug')" });
+            ec.Processors.Add(new WorkItemMigrationConfig() { Enabled = false, ReplayRevisions= true, LinkMigration = true, UpdateCreatedBy =true, PrefixProjectToNodes = false, UpdateCreatedDate=true, UpdateSourceReflectedId=true, QueryBit = @"AND [TfsMigrationTool.ReflectedWorkItemId] = '' AND  [Microsoft.VSTS.Common.ClosedDate] = '' AND [System.WorkItemType] IN ('Shared Steps', 'Shared Parameter', 'Test Case', 'Requirement', 'Task', 'User Story', 'Bug')" });
             ec.Processors.Add(new WorkItemUpdateConfig() { Enabled = false, QueryBit = @"AND [TfsMigrationTool.ReflectedWorkItemId] = '' AND  [Microsoft.VSTS.Common.ClosedDate] = '' AND [System.WorkItemType] IN ('Shared Steps', 'Shared Parameter', 'Test Case', 'Requirement', 'Task', 'User Story', 'Bug')" });
             ec.Processors.Add(new NodeStructuresMigrationConfig() { Enabled = false, PrefixProjectToNodes=false, BasePaths = new [] { "Product\\Area\\Path1", "Product\\Area\\Path2" } });
             ec.Processors.Add(new LinkMigrationConfig() { Enabled = false, QueryBit= @"AND ([System.ExternalLinkCount] > 0 OR [System.RelatedLinkCount] > 0)" });
